+++ src/Plugin/Field/FieldWidget/Select2EntityReference.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace Drupal\select2\Plugin\Field\FieldWidget;
+
+use Drupal\Core\Field\FieldItemListInterface;
+use Drupal\Core\Form\FormStateInterface;
+
+/**
+ * Plugin implementation of the 'select2' widget.
+ *
+ * @FieldWidget(
+ *   id = "select2_entity_reference",
+ *   label = @Translation("Select2 (simple select)"),
+ *   field_types = {
+ *     "entity_reference",
+ *   },
+ *   multiple_values = TRUE
+ * )
+ */
+class Select2EntityReference extends Select2Widget {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function settingsForm(array $form, FormStateInterface $form_state) {
+    $element = parent::settingsForm($form, $form_state);
+    $element['autocomplete'] = [
+      '#type' => 'checkbox',
+      '#title' => t('Autocomplete'),
+      '#default_value' => $this->getSetting('autocomplete'),
+      '#description' => t('Options will be lazy loaded. This is recommended for lists with a lot of values.'),
+    ];
+
+    return $element;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function settingsSummary() {
+    $summary = parent::settingsSummary();
+    $autocomplete = $this->getSetting('autocomplete');
+    $summary[] = t('Autocomplete: @autocomplete', ['@autocomplete' => $autocomplete ? $this->t('On') : $this->t('Off')]);
+    return $summary;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state) {
+    $element = parent::formElement($items, $delta, $element, $form, $form_state);
+
+    if ($this->getSetting('autocomplete')) {
+      $element['#autocomplete'] = TRUE;
+      $element['#target_type'] = $this->getFieldSetting('target_type');
+      $element['#selection_handler'] = $this->getFieldSetting('handler');
+      $element['#selection_settings'] = $this->getFieldSetting('handler_settings') + ['match_operator' => 'CONTAINS'];
+    }
+
+    return $element;
+  }
+
+}
