+++ src/Plugin/Field/FieldWidget/Select2Widget.php
@@ -13,7 +13,6 @@ use Drupal\Core\Form\FormStateInterface;
  *   id = "select2",
  *   label = @Translation("Select2 (simple select)"),
  *   field_types = {
- *     "entity_reference",
  *     "list_integer",
  *     "list_float",
  *     "list_string"
@@ -35,45 +34,9 @@ class Select2Widget extends OptionsSelectWidget {
   /**
    * {@inheritdoc}
    */
-  public function settingsForm(array $form, FormStateInterface $form_state) {
-    $element = [];
-    if ($this->fieldDefinition->getType() == 'entity_reference') {
-      $element['autocomplete'] = [
-        '#type' => 'checkbox',
-        '#title' => t('Autocomplete'),
-        '#default_value' => $this->getSetting('autocomplete'),
-        '#description' => t('Options will be lazy loaded. This is recommended for lists with a lot of values.'),
-      ];
-    }
-
-    return $element;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function settingsSummary() {
-    $summary = [];
-    if ($this->fieldDefinition->getType() == 'entity_reference') {
-      $autocomplete = $this->getSetting('autocomplete');
-      $summary[] = t('Autocomplete: @autocomplete', ['@autocomplete' => $autocomplete ? $this->t('On') : $this->t('Off')]);
-    }
-    return $summary;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
   public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state) {
     $element = parent::formElement($items, $delta, $element, $form, $form_state);
     $element['#type'] = 'select2';
-
-    if ($this->getSetting('autocomplete')) {
-      $element['#autocomplete'] = TRUE;
-      $element['#target_type'] = $this->getFieldSetting('target_type');
-      $element['#selection_handler'] = $this->getFieldSetting('handler');
-      $element['#selection_settings'] = $this->getFieldSetting('handler_settings') + ['match_operator' => 'CONTAINS'];
-    }
 
     return $element;
   }
