+++ tests/src/Kernel/Element/Select2Test.php
@@ -22,14 +22,14 @@ class Select2Test extends KernelTestBase {
   public function testSelect2Theming() {
     $select = [
       '#type' => 'select2',
-      '#options' => ['Foo', 'Bar'],
+      '#options' => [],
     ];
 
     $this->render($select);
-
-    $select2_js = $this->xpath("//script[contains(@src, 'select2.js')]");
+    $select2_js = $this->xpath("//script[contains(@src, 'modules/select2/js/select2.js')]");
     $this->assertEqual(count($select2_js), 1);
-
+    $select2_js = $this->xpath("//script[contains(@src, 'select2/dist/js/select2.min.js')]");
+    $this->assertEqual(count($select2_js), 1);
   }
 
 }
